chore(windows): move systeminfo helper in an internal package that support build constraints
